% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_sample.R
\docType{data}
\name{FacetSample}
\alias{FacetSample}
\title{The FacetSample object.}
\format{An object of class \code{FacetSample} (inherits from \code{FacetGrid}, \code{Facet}, \code{ggproto}, \code{gg}) of length 9.}
\description{
refer to \code{\link[ggplot2]{Facet}}
}
\note{
The general idea is to generate a data-frame for each facet dimension,
with a unique set of keys (initially all possible combinations) for each dimension,
then sample without replacement from that unique-key list,
then left-join that key list to the original data, leaving only the relevant data to display.
}
\keyword{datasets}
